simplepack: Only clean up temp dir in first process in case of fork
If a process forks, all the children attempt to execute the atexit
handlers. This causes the first child to exit to delete the directory
containing all the unzipped resources, which causes errors later. To
avoid this, record the pid when we create the tempdir, and only clean
up in that process.