simplepack: Add a horrific hack to make pkg_resources work in zip
google.cloud.datastore imports pkg_resources to use get_distribution
to pull out the version number. Unfortunately, if sys.path contains
the directory with the zip, pkg_resources will decide it should put
the zip at the beginning of sys.path, which screws up our carefully
ordered imports.

To avoid this, hack a constant in the module which changes the
precedence, effectively disabling this behavior.