pyz_library: force_unzip should apply to wheels as well as srcs
The certifi PyPI package contains .pem files, and exports the path to
them, relative to __file__. This means it must be unzipped. The
force_unzip attribute should do this, but previously it only did this
for srcs and data. This change makes it apply to wheels, by passing
the path of the wheel to simplepack, which then expands it to the list
of files in that wheel.

pip_generate also has a hack added to add this attribute for certifi.

README.md: Update example commit; fix typo
pip_generate.go: Add map of PyPI packages that must be unzipped
rules_python_zip.bzl: Add wheel files to force_unzips
simplepack.go:
* Close zip readers
* Merge contents of wheels for force unzips
tests: test certifi to verify that force_unzip works with wheels
tools: Regenerate binaries