pyz_binary: Include setuptools if it will be executed in a zip
setuptools includes pkg_resources, which is used to load resources
from a zip. This allows packages like pytz to be run from inside a
zip. Change pyz_binary to detect if it is needed, and add it if so.

Rename all the pypi libraries we depend on to pypi_* to match the
paths generated by pip_generate, and add them conditionally. This
should allow these rules to co-exist with packages that directly
depend on setuptools.