simplepack: re-exec with python -ESs to avoid external environment
This attempts to ensure we don't run with anything except the standard
library. At Bluecore, we used .pth files to add App Engine imports to the
Python environment by default. However, this causes version conflicts with
targets built with Bazel, which bundle their own version of the App Engine
Python libraries.

To resolve it, run with python -S. However, Pythons created by virtualenv
don't have the runpy module needed to execute .zip files. As a result, we
need to re-execute ourselves with the correct PYTHONPATH to make it work.

Copy the initialization from virtualenv to set up the environment, if
running inside a virtualenv.

virtualenv_imports_test: Test with a custom .pth