simplepack: Unset PYTHONPATH so subprocesses behave correctly
Commit fec3053469 made the pyz enviroment isolated from external Python
virtualenvs and environment variables. However, it used PYTHONPATH to
re-execute the binary without site-packages, but with the runpy module.
This broke virtualenv and our tests on Linux, since it ended up finding
the incorrect settings module in the subprocesses that were spawned.

To fix it, remove PYTHONPATH so the startup script has full control over
the Python import paths.