pip_generate.go: re-add wheelURLPrefix (#5)
Commit 951d8afe56 removed this feature, which is needed by Bluecore,
since we have a developer run pip_generate and we do not check in the
compiled wheels. Add it back.

wheelsDir: Make relative to CWD not outputDir: We need to generate wheels
    outside the Bazel workspace.
rulesWorkspace: Default to com_bluecore_rules_pyz which matches the
    convention used by the "official" rules (io_bazel_rules_foo)

Call native.existing_rules() once to match "best practices".